Table can now scroll
