+++ eq-author/src/components/modals/QuestionnaireSelectModal/index.js
@@ -1,11 +1,15 @@
-import React from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
+
+import * as Headings from "constants/table-headings";
 
 import styled from "styled-components";
 import { colors } from "constants/theme";
+
 import Modal from "components/modals/Modal";
 import Button from "components/buttons/Button";
 import ButtonGroup from "components/buttons/ButtonGroup";
+import QuestionnairesView from "components/QuestionnairesView";
 
 const ModalFooter = styled.footer`
   padding: 1.5em;
@@ -31,10 +35,6 @@ const ModalTitle = styled.h2`
   padding-left: 1em;
 `;
 
-const Panel = styled.div`
-  background-color: ${colors.white};
-`;
-
 const MenuContainer = styled.div`
   overflow: hidden;
   height: 28em;
@@ -44,18 +44,36 @@ const QuestionnaireSelectModal = ({
   isOpen,
   onClose,
   onSelect,
-  children,
-  disableSelect,
+  questionnaires,
 }) => {
+  const [selectedQuestionnaire, setSelectedQuestionnaire] = useState(null);
+
+  const enabledHeadings = [
+    Headings.TITLE,
+    Headings.OWNER,
+    Headings.CREATED,
+    Headings.MODIFIED,
+  ];
+
   return (
     <StyledModal isOpen={isOpen} onClose={onClose} hasCloseButton>
       <Container data-test="questionnaire-select-modal">
-        <>
-          <MenuContainer>
-            <ModalTitle>Select the source questionnaire</ModalTitle>
-            <Panel>{children}</Panel>
-          </MenuContainer>
-        </>
+        <MenuContainer>
+          <ModalTitle>Select the source questionnaire</ModalTitle>
+          <QuestionnairesView
+            questionnaires={questionnaires}
+            selectedQuestionnaire={selectedQuestionnaire}
+            enabledHeadings={enabledHeadings}
+            canCreateQuestionnaire={false}
+            padding="small"
+            variant="selectModal"
+            onQuestionnaireClick={(questionnaireId) =>
+              setSelectedQuestionnaire(
+                questionnaires.find(({ id }) => id === questionnaireId)
+              )
+            }
+          />
+        </MenuContainer>
       </Container>
       <ModalFooter>
         <ButtonGroup horizontal align="right">
@@ -64,9 +82,8 @@ const QuestionnaireSelectModal = ({
           </Button>
           <Button
             variant="primary"
-            autoFocus
-            onClick={onSelect}
-            disabled={disableSelect}
+            onClick={() => onSelect(selectedQuestionnaire)}
+            disabled={!selectedQuestionnaire}
           >
             Select
           </Button>
